Update tests to validate the DICOM format
The switch from RFC3881 to DICOM required a rethink of the testing code.
Now the tests require xmllint to be on the path and uses that for DICOM
schema testing. This is required as Relax NG is used for the DICOM
schema and there aren't any good tools in js to do Relax NG validation.